+++ NEWS.md
@@ -5,9 +5,13 @@ arg News and Updates
 
 * Added `arg_color()` (and variants) to check whether an argument corresponds to a valid color specification.
 
-* Fixed a bug where the message provided to `.msg` (if any) was not evaluated in the correct environment.
+* Added `arg_name()` and `arg_names()` to check whether an argument is a vector of valid names of elements of an object.
+
+* `arg_index()` and `arg_indices()` now check names for arbitrary objects, not just matrices and data frames.
 
 * `arg_counts()` and `arg_whole_numeric()` (and their scalar equivalents) now check whether a value is a whole number using `rlang::is_integerish()` instead of the criterion used by `all.equal()`.
+
+* Fixed a bug where the message provided to `.msg` (if any) was not evaluated in the correct environment.
 
 # *arg* 0.1.1